rename command group jcoin to frictionless
